refactor(attribution): fold embedCode into baseBuilderCode, drop stale aggregator reference
The bc_julg9gbq builder code was reverted out of the aggregator, so the
'must match aggregator/services/builder_code.go' note no longer points at
anything. Replaced with what the code actually is.

- merge app/lib/embedCode.ts into app/lib/baseBuilderCode.ts
- embedCodeToHex now delegates to the existing stringToHex helper
- consolidate the three attribution test files into baseBuilderCode.test.ts
- remove EMBED_ATTRIBUTION_IMPLEMENTATION.md, superseded by docs/embed-widget.md;
  its warning about rebuilding multi-code suffixes moved into a code comment
